+++ src/core/json/analysis/public.cpp
@@ -7,6 +7,7 @@ namespace json {
 
     void write_analysis_variables_to_json(const io::MatlabContext& context, const std::string& data_dir_path) {
         nlohmann::json variables = {
+            {"aircraft_id", context.aircraft_id},
             {"state_dim", context.state_dim},
             {"input_dim", context.input_dim},
             {"matrix_paths", {
